refactor(core): make core own dialect implementation
