+++ crates/copybook-core/Cargo.toml
@@ -26,7 +26,6 @@ copybook-error.workspace = true
 copybook-error-reporter.workspace = true
 copybook-overflow.workspace = true
 copybook-utils.workspace = true
-copybook-dialect.workspace = true
 copybook-lexer.workspace = true
 serde.workspace = true
 serde_json.workspace = true
