+++ crates/copybook-dialect/Cargo.toml
@@ -21,10 +21,7 @@ all-features = false
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-serde.workspace = true
-
-[dev-dependencies]
-serde_json.workspace = true
+copybook-core.workspace = true
 
 [features]
 default = []
