+++ crates/copybook-dialect/README.md
@@ -1,12 +1,13 @@
 # copybook-dialect
 
-Shared dialect contract for ODO (`OCCURS DEPENDING ON`) `min_count` semantics.
+Deprecated compatibility package for the ODO (`OCCURS DEPENDING ON`) dialect
+contract.
 
 ## Overview
 
-Different COBOL compilers interpret the minimum occurrence count in ODO declarations
-differently. This crate provides the `Dialect` enum and an `effective_min_count` helper
-that normalizes declared bounds according to the selected dialect (IBM, Micro Focus, etc.).
+The implementation now lives in `copybook-core::dialect`. This package forwards
+the 0.5 API during the 0.6 compatibility window and will not receive new
+implementation behavior.
 
 ## Usage
 
@@ -19,7 +20,15 @@ assert_eq!(effective_min_count(dialect, 0), 1);
 # Ok::<(), String>(())
 ```
 
-## Public API
+## Migration
+
+Use the core-owned path for new code:
+
+```rust
+use copybook_core::dialect::{Dialect, effective_min_count};
+```
+
+## Compatibility API
 
 - `Dialect` — `Normative`, `ZeroTolerant`, `OneTolerant`
 - `effective_min_count(dialect, declared_min_count)` — Apply dialect rules
