+++ crates/copybook-dialect/src/claude.md
@@ -2,7 +2,7 @@
 # claude.md
 
 ## Scope
-- Library sources for `copybook-dialect` (Schema layer).
+- Deprecated forwarding source for the core-owned dialect contract.
 
 ## Navigation
 - Crate root: `../claude.md`
