+++ crates/copybook-dialect/tests/claude.md
@@ -2,7 +2,7 @@
 # claude.md
 
 ## Scope
-- Integration tests for `copybook-dialect` (Schema layer).
+- Compatibility compile-contract tests for `copybook-dialect`.
 
 ## Navigation
 - Crate root: `../claude.md`
